TeamCalendar funcional.

+++ nicolaspuebla_proyecto_final_android/app/src/main/java/com/example/nicolaspuebla_proyecto_final_android/data/model/dataClases/Event.kt
@@ -6,7 +6,7 @@ open class Event(
     open val id: Long,
     open val team: Team,
     open val address: String,
-    open val latitude: Int,
-    open val longitude: Int,
+    open val latitude: Double,
+    open val longitude: Double,
     open val date: String
 )
