+++ nicolaspuebla_proyecto_final_android/app/src/main/java/com/example/nicolaspuebla_proyecto_final_android/data/model/dataClases/Training.kt
@@ -6,7 +6,7 @@ class Training(
     override val id: Long,
     override val team: Team,
     override val address: String,
-    override val latitude: Int,
-    override val longitude: Int,
+    override val latitude: Double,
+    override val longitude: Double,
     override val date: String,
 ) : Event(id, team, address, latitude, longitude, date)
